+++ sdk/storage/Microsoft.Azure.WebJobs.Extensions.Storage.Blobs/src/Bindings/IBlobCommitedAction.cs
@@ -4,7 +4,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Microsoft.Azure.WebJobs.Host.Blobs.Bindings
+namespace Microsoft.Azure.WebJobs.Extensions.Storage.Blobs.Bindings
 {
     internal interface IBlobCommitedAction
     {